Modularize code in cleaning.py
