+++ life_expectancy/__init__.py
@@ -1,0 +1,4 @@
+"init file"
+from .main import main
+
+__all__ = ["main"]
